+++ test/learningAgentTest.cpp
@@ -124,6 +124,16 @@ TEST_F(LearningAgentTest, Init)
         << "Initialization of the LearningAgent should not fail.";
 }
 
+TEST_F(LearningAgentTest, InitNbRoots)
+{
+    params.mutation.tpg.initNbRoots = 42;
+    Learn::LearningAgent la(le, set, params);
+
+    ASSERT_NO_THROW(la.getTPGGraph()->getNbRootVertices() != 42)
+        << "Initialization of the LearningAgent should have a number of roots equal to the number specify";
+}
+
+
 TEST_F(LearningAgentTest, addLogger)
 {
     params.archiveSize = 50;
@@ -785,6 +795,8 @@ TEST_F(ParallelLearningAgentTest, EvalRootSequential)
     // Initialize Randomness
     Mutator::RNG rng;
     rng.setSeed(0);
+
+    params.mutation.tpg.initNbRoots = params.mutation.tpg.nbActions;
 
     // Initialize the tpg
     Mutator::TPGMutator::initRandomTPG(tpg, params.mutation, rng);